+++ src/src/zuno/config.js
@@ -28,6 +28,14 @@ const _this = {//У нас активируеться только когда п
 	{
 		_set(ZunoConstant.PATH.JSON_WORKSPACE, 'security', value);
 	},
+	getPower()//Получает используемую при загрузки скетча мощность передаваемого радио сигнала из настроект текущего проекта
+	{
+		return (_arrayFind(ZunoConstant.POWER, _get(ZunoConstant.PATH.JSON_WORKSPACE, 'power'), ZunoConstant.POWER_DEFAULT));
+	},
+	setPower(value)//Сохраняет используемый скетч из настроект текущего проекта
+	{
+		_set(ZunoConstant.PATH.JSON_WORKSPACE, 'power', value);
+	},
 	getFrequency()//Получает используемую при загрузки скетча частоту из настроект текущего проекта
 	{
 		return (_arrayFind(ZunoConstant.FREQUENCY, _get(ZunoConstant.PATH.JSON_WORKSPACE, 'frequency'), ZunoConstant.FREQUENCY_DEFAULT));